refactor(scripts): consolidate releaser into scripts/lib
- Move targets to scripts/lib/targets.ts (shared by build, sign-macos,
  and releaser)
- Move run() and publish() helpers into scripts/lib/npm.ts alongside
  the existing isPublished()
- Flatten scripts/releaser/index.ts to scripts/releaser.ts now that
  the directory only held one file
- Update import paths in build.ts, sign-macos.ts, and package.json
