+++ package.json
@@ -19,9 +19,9 @@
     "version-packages": "bun changeset version",
     "version-packages:canary": "bun run scripts/snapshot.ts -- canary",
     "version-packages:snapshot": "bun run scripts/snapshot.ts",
-    "release": "bun run scripts/releaser/index.ts",
-    "release:canary": "bun run scripts/releaser/index.ts --tag canary",
-    "release:snapshot": "bun run scripts/releaser/index.ts --tag snapshot",
+    "release": "bun run scripts/releaser.ts",
+    "release:canary": "bun run scripts/releaser.ts --tag canary",
+    "release:snapshot": "bun run scripts/releaser.ts --tag snapshot",
     "build:compile:all": "bun run scripts/build.ts",
     "global:link": "bun link --cwd packages/cli-core",
     "global:unlink": "bun unlink --cwd packages/cli-core",
